Extract poll_loop to another file
